Deploy 17 - refactor x errores de DB

+++ src/main/java/com/psa/soporte/DTO/request/TicketRequest.java
@@ -26,5 +26,5 @@ public class TicketRequest {
     @Schema(example = "0")
     private Integer colaboradorId;
     @Schema(example = "[1,2,3,4,5]")
-    private List<Integer> tareaIds;
+    private List<Long> tareaIds;
 }
